resolving conflicts and merging to refactor
