+++ src/main/java/com/sprint/be_java_hisp_w23_g04/gateways/IUserGateway.java
@@ -9,5 +9,4 @@ public interface IUserGateway {
     List<User> getByIds(List<Integer> usersId);
     List<User> getAllUsers();
     void unfollowUser(int userId, int unfollowId);
-    List<User> findAll();
 }
